+++ app/src/main/res/layout/fragment_list.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListFragment">
+    tools:context=".ui.ListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_book"
